fix delete project issue
